src/requirements.txt
numpy
pandas
matplotlib
fsspec
pyyaml